move JoinedStrNode cleanup into ConstantFolding transform as that's essentially what it does
